move frame_sync into it's own module
because:
- it doesn't need to all be in one huge file
